Adding Tests

+++ Google_tests/lib/googletest/include/gtest/gtest.h
@@ -1375,7 +1375,7 @@ template <typename T1, typename T2>
 AssertionResult CmpHelperEQ(const char* lhs_expression,
                             const char* rhs_expression, const T1& lhs,
                             const T2& rhs) {
-  if (static_cast<const int>(lhs) == rhs) {
+  if ((lhs) == rhs) {
     return AssertionSuccess();
   }
